Fix AI code on 64-bit systems
Implemented from Debian patch: https://sources.debian.org/src/gltron/0.70final-14/debian/patches/amd64-ai.patch

+++ src/include/game/game_data.h
@@ -73,7 +73,7 @@ typedef struct Data {
 typedef struct AI {
 	int active;
 	int tdiff;
-	long lasttime;
+	unsigned int lasttime;
 	segment2 left, right, front, backleft;
 } AI;
